MNT/FIX: Only build Lambda Layers and use Lambdas with x86 architecture (#496)
* MNT/FIX: Only build Lambda Layers and use Lambdas with x86 architecture

This was causing issues where people with Mac / ARM processors would
build a different layer than on GitHub Actions CI on a Linux x86
system. We can specify the platform for the Layer explicitly and
indicate what it is compatible with. Then we can remove all ARM64
architectures for the Lambda's we were specifying which aren't
needed anymore if we build the layer as an x86.

* DOC: Add documentation for how to debug the cdk deploy for architectures.