MNT: Delay packet downloader cron job
The POC has indicated that there could be a delay between when the
file arrives and when it gets ingested and available in the database.
I think this means there could still potentially be a further delay
so lets not push it too close to the 0/6/12/18 boundaries and give
it 20 minutes buffer to fill the database after each request.